G2-1546 Test SSH w/port forwarding
